pdf modules refactor
