+++ backend/main.py
@@ -56,8 +56,8 @@ from api.messages import router as messages_router
 from api.student.notifications import router as student_notifications_router
 from api.student.cohorts import router as student_cohorts_router
 from api.student.simulation_instances import router as student_simulation_instances_router
-from api.parse_pdf import router as pdf_router
-from api.pdf_progress import router as progress_router
+# PDF processing now uses modular structure
+from modules.pdf_processing.router import router as pdf_router
 from api.simulation import router as simulation_router
 from api.publishing import router as publishing_router
 from api.oauth import router as oauth_router, lifespan as oauth_lifespan
@@ -262,8 +262,8 @@ async def global_exception_handler(request: Request, exc: Exception):
 # Session activity middleware removed - using JWT token expiration instead
 
 # Include API routers
-app.include_router(pdf_router, tags=["PDF Processing"])
-app.include_router(progress_router, tags=["PDF Progress"])
+# PDF processing router now includes both PDF parsing AND progress endpoints (no separate progress_router)
+app.include_router(pdf_router, tags=["PDF Processing"])  # Now using modular structure
 app.include_router(simulation_router, tags=["Simulation"])
 app.include_router(publishing_router, tags=["Publishing"])
 app.include_router(oauth_router, tags=["OAuth"])
@@ -278,8 +278,8 @@ app.include_router(student_notifications_router, tags=["Student Notifications"])
 app.include_router(student_cohorts_router, tags=["Student Cohorts"])
 app.include_router(student_simulation_instances_router, tags=["Student Simulation Instances"])
 
-# Import progress manager for WebSocket endpoint
-from api.pdf_progress import progress_manager
+# Import progress manager for WebSocket endpoint from new module
+from modules.pdf_processing.progress_service import progress_manager
 
 # Add WebSocket endpoint directly to the app
 @app.websocket("/ws/pdf-progress/{session_id}")
