+++ backend/tests/modules/pdf_processing/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tests for PDF processing module.
+"""
